finish to correct mediadb problem
darcs-hash:20070504181904-73b03-1731d99f09114d3faf7c8fb81454ee9287e9bd02.gz

committer: Mickael ROYER <mickael.royer@gmail.com>

+++ testdeejayd/databuilder.py
@@ -165,11 +165,12 @@ class TestMediaCollection(TestProvidedMusicCollection):
 
     def __init__(self):
         self.dir_struct_written = False
+        self.clean_library = True
         self.dirs = {} 
         self.medias = {} 
 
     def cleanLibraryDirectoryTree(self):
-        if self.dir_struct_written:
+        if self.dir_struct_written and self.clean_library:
             shutil.rmtree(self.datadir)
 
     def buildMusicDirectoryTree(self, destDir = "/tmp"):
